requirements.txt
sentence_transformers
pandas
opensearch-py
python-dotenv